Moved Text and Drawing out
